Move paramselectionstate to its own file
